SplitContainers: Make Twitch live/mentions/etc tabs 'sticky'
Following on from #4358 (add setting to only show tabs with
'Live' channels) I noted that the channel types for 'live',
'mentions', 'watches', and 'whispers' disappear as they are
not truly online channels. This commit adds a check to set
the 'online' flag on tabs containing these channel types so
that they always appear in the tab list/notebook.

A future improvement/commit will probably add a further
settings option to toggle this behaviour on/off if desired.

Signed-off-by: Michael Everitt <michael@2e0cer.net>